requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==4.7.0
azure-storage-blob==12.21.0
beautifulsoup4==4.12.3
blinker==1.8.2
botbuilder-core==4.16.1
botbuilder-integration-aiohttp==4.16.1
botbuilder-schema==4.16.1
botframework-connector==4.16.1
botframework-streaming==4.16.1
boto3==1.34.145
botocore==1.34.145
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.6.1
colorama==0.4.6
cryptography==42.0.8
dataclasses-json==0.6.7
distro==1.9.0
docx2txt==0.8
exceptiongroup==1.2.2
fastapi==0.110.3
fastavro==1.9.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpickle==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.10
langchain-cli==0.0.25
langchain-cohere==0.1.5
langchain-community==0.2.9
langchain-core==0.2.22
langchain-experimental==0.0.62
langchain-mistralai==0.1.10
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langdetect==1.0.9
langgraph==0.1.9
langserve==0.2.2
langsmith==0.1.93
libcst==1.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
msal==1.30.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.36.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.4.0
protobuf==5.27.2
pyarrow==17.0.0
pycparser==2.22
pydantic==1.10.13
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyodbc==5.1.0
pypdf==4.3.0
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sse-starlette==1.8.2
starlette==0.37.2
streamlit==1.36.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.9.4
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
uvicorn==0.23.2
watchdog==4.0.1
yarl==1.9.4
